Initial commit of moving first Italic check from GF to OT
