+++ Lib/fontbakery/profiles/googlefonts.py
@@ -18,6 +18,7 @@ from fontbakery.constants import (NameID,
                                   MacintoshLanguageID,
                                   LATEST_TTFAUTOHINT_VERSION)
 from .googlefonts_conditions import * # pylint: disable=wildcard-import,unused-wildcard-import
+from .shared_conditions import style
 from glyphsets import codepoints
 ENCODINGS_DIR = codepoints.nam_dir
 
@@ -34,6 +35,7 @@ profile.configuration_defaults = {
 
 OVERRIDDEN_CHECKS = [
     "com.adobe.fonts/check/freetype_rasterizer",
+    "com.google.fonts/check/italic_angle",
 ]
 
 METADATA_CHECKS = [
@@ -157,7 +159,6 @@ FONT_FILE_CHECKS = [
     'com.google.fonts/check/version_bump',
     'com.google.fonts/check/epar',
     'com.google.fonts/check/font_copyright',
-    'com.google.fonts/check/italic_angle',
     'com.google.fonts/check/slant_direction',
     'com.google.fonts/check/has_ttfautohint_params',
     'com.google.fonts/check/name/version_format',
@@ -3127,69 +3128,6 @@ def com_google_fonts_check_fsselection(ttFont, style):
                           expected,
                           bitmask=FsSelection.BOLD,
                           bitname="BOLD")
-
-
-@check(
-    id = 'com.google.fonts/check/italic_angle',
-    conditions = ['style'],
-    rationale = """
-        The 'post' table italicAngle property should be a reasonable amount, likely
-        not more than -20°, never more than -30°, and never greater than 0°. Note that
-        in the OpenType specification, the value is negative for a lean rightwards.
-
-        https://docs.microsoft.com/en-us/typography/opentype/spec/post
-    """,
-    proposal = 'legacy:check/130'
-)
-def com_google_fonts_check_italic_angle(ttFont, style):
-    """Checking post.italicAngle value."""
-    failed = False
-    value = ttFont["post"].italicAngle
-
-    # Checking that italicAngle <= 0
-    if value > 0:
-        failed = True
-        yield FAIL,\
-              Message("positive",
-                      (f"The value of post.italicAngle is positive, which"
-                       f" is likely a mistake and should become negative,"
-                       f" from {value} to {-value}."))
-
-    # Checking that italicAngle is less than 20° (not good) or 30° (bad)
-    # Also note we invert the value to check it in a clear way
-    if abs(value) > 30:
-        failed = True
-        yield FAIL,\
-              Message("over-minus30-degrees",
-                      (f"The value of post.italicAngle ({value}) is very high"
-                       f" (over -30°!) and should be confirmed."))
-    elif abs(value) > 20:
-        failed = True
-        yield WARN,\
-              Message("over-minus20-degrees",
-                      (f"The value of post.italicAngle ({value}) seems very high"
-                       f" (over -20°!) and should be confirmed."))
-
-
-    # Checking if italicAngle matches font style:
-    if "Italic" in style:
-        if ttFont['post'].italicAngle == 0:
-            failed = True
-            yield FAIL,\
-                  Message("zero-italic",
-                          ("Font is italic, so post.italicAngle"
-                           " should be non-zero."))
-    else:
-        if ttFont["post"].italicAngle != 0:
-            failed = True
-            yield FAIL,\
-                  Message("non-zero-normal",
-                          ("Font is not italic, so post.italicAngle"
-                           " should be equal to zero."))
-
-    if not failed:
-        yield PASS, (f'Value of post.italicAngle is {value}'
-                     f' with style="{style}".')
 
 
 @condition
@@ -6488,6 +6426,13 @@ profile.check_log_override(
     "com.adobe.fonts/check/freetype_rasterizer",
     overrides=(("freetype-not-installed", FAIL, KEEP_ORIGINAL_MESSAGE),),
     reason="For Google Fonts, this check is very important and should never be skipped.",
+)
+
+profile.check_log_override(
+    # From opentype.py
+    "com.google.fonts/check/italic_angle",
+    overrides=(("over-minus30-degrees", FAIL, KEEP_ORIGINAL_MESSAGE),),
+    reason=("For Google Fonts, an Italic angle over -30° is considered a FAIL."),
 )
 
 GOOGLEFONTS_PROFILE_CHECKS = add_check_overrides(
