+++ Lib/fontbakery/profiles/opentype.py
@@ -75,6 +75,7 @@ OPENTYPE_PROFILE_CHECKS = [
     'com.google.fonts/check/layout_valid_feature_tags',
     'com.google.fonts/check/layout_valid_script_tags',
     'com.google.fonts/check/layout_valid_language_tags',
+    'com.google.fonts/check/italic_angle',
     'com.adobe.fonts/check/varfont/valid_axis_nameid',
     'com.adobe.fonts/check/varfont/valid_subfamily_nameid',
     'com.adobe.fonts/check/varfont/valid_postscript_nameid',
